refactor: move test files to top-level tests/ directory
