+++ tests/destinations/slack.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi } from 'vitest'
 
-import { Alert } from '../entities'
+import { Alert } from '../../src/entities'
 
-import { sendAlertsToSlack, validateSlackWebhookUrl } from './slack'
+import { sendAlertsToSlack, validateSlackWebhookUrl } from '../../src/destinations/slack'
 
 // Split to avoid triggering secret-scanning heuristics on the hooks.slack.com pattern
 const SLACK_BASE = 'https://hooks' + '.slack.com/services'
