+++ tests/entities/vulnerability.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 
-import { toVulnerability } from './vulnerability'
+import { toVulnerability } from '../../src/entities/vulnerability'
 
 describe('toVulnerability', () => {
   it('maps all fields correctly', () => {
