+++ vitest.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   test: {
     clearMocks: true,
     environment: 'node',
-    include: ['src/**/*.test.ts'],
+    include: ['tests/**/*.test.ts'],
   },
 })
